v2

+++ app.py
@@ -4,7 +4,7 @@ import plotly.express as px
 
 
 df = pd.read_csv('vehicles_us.csv')
-df['price'] = df['price'].astype(float)
+df['price'] = df['price'].astype(int)
 
 df['manufacturer'] = df['model'].apply(lambda x: x.split()[0])
